ConstantBuffer wrapper. Beginning of Camera. Sprite Rendering. Mouse Wheel.
Added a wrapper for ContantBuffers to make it easier to set data to
shaders. Added CameraData and TimeData as examples.

Created a Camera class and sent its data to the shaders using the new
ConstantBuffer class.

Once CSprite can now be used to draw more sprites on the screen by
changing its data between the Render calls.

Added mouse wheel input to the InputManager.